smart-city-allocation/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic==2.10.3
pydantic-settings==2.6.1
joblib==1.4.2
pandas==2.2.3
scikit-learn==1.5.2
numpy==1.26.4
xgboost==2.1.3
matplotlib==3.9.2
seaborn==0.13.2
jupyter==1.1.1
python-multipart==0.0.17
gunicorn==23.0.0
requests==2.32.3
shap==0.46.0
python-dotenv==1.0.1
sqlalchemy==2.0.36
alembic==1.14.0
slowapi==0.1.9
websockets==14.1
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
